% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs-use-most.R
\name{pkg_use}
\alias{pkg_use}
\title{Package use}
\usage{
pkg_use(...)
}
\arguments{
\item{...}{Directory location in which files ending in '.R' and '.Rmd' will
be searched for packages (via fully qualified namespace or calls to library
or require).}
}
\value{
A pkguse object, which is really just a tibble with specialized
  methods for plot and summary.
}
\description{
Take inventory of your own package use
}
